move providerId to types/lists
