+++ src/contexts/WalletContext/wallet-connect/types.ts
@@ -1,4 +1,4 @@
-import { ProviderId } from "../types";
+import { ProviderId } from "types/lists";
 
 export type WalletMeta = { logo: string; name: string; providerId: ProviderId };
 type Loading = { status: "loading" };
