+++ src/types/lists.ts
@@ -19,3 +19,15 @@ export type UNSDG_NUMS =
   | 15
   | 16
   | 17;
+
+export type ProviderId =
+  | "binance-wallet"
+  | "metamask"
+  | "evm-wc"
+  | "xdefi-wallet" //xdefi terra provider
+  | "xdefi-evm" //xdefi evm provider
+  | "leap-wallet"
+  | "station"
+  | "walletconnect"
+  | "keplr-wc"
+  | "keplr";
